test(plugins): rename test-only helper files
