+++ extensions/slack/test-api.ts
@@ -3,7 +3,7 @@ export type { SlackMessageEvent } from "./src/types.js";
 export { slackPlugin } from "./src/channel.js";
 export { setSlackRuntime } from "./src/runtime.js";
 export { createSlackActions } from "./src/channel-actions.js";
-export { createSlackOutboundPayloadHarness } from "./src/outbound-payload-harness.js";
+export { createSlackOutboundPayloadHarness } from "./src/outbound-payload.test-harness.js";
 export { prepareSlackMessage } from "./src/monitor/message-handler/prepare.js";
 export { createInboundSlackTestContext } from "./src/monitor/message-handler/prepare.test-helpers.js";
 export { slackOutbound } from "./src/outbound-adapter.js";
